Readme.md
In cellular automata, the von Neumann neighborhood (or 4-neighborhood) is classically defined on a two-dimensional square lattice and is composed of a central cell and its four adjacent cells. 

In cellular automata, the Moore neighborhood is defined on a two-dimensional square lattice and is composed of a central cell and the eight cells that surround it.

I used Moore and Neumann neighborhood models to predict COVID-19 dynamics.  

In this project, I have performed different simulations as follows:

1. Using the Neumann Neighborhood model to predict Susceptible, Infected, Recovered, and Cumulative population as a function of time.
2. Using the Moore Neighborhood model to predict Susceptible, Infected, Recovered, and Cumulative population as a function of time.
# I compared the results from these models to the statistics provided by the Centre for Disease Control (CDC) to determine the accuracy of the models.

3. I used Cellular Automata model to determine how social distancing affected the cumulative infection rate.